Add files via upload
This is the Channel Strip for the VCM600 Script. Features DJ EQ device and Filter device just like on the 6 Tracks on the Controller. The Script assigns the 6 first tracks to the 6 channels on the controller. The best is to have your first 6 tracks as Audio busses to which you route everything. No track scrolling implemented. 6 tracks max. period.